backend/ai_tutor/requirements.txt
openai>=1.55.3
pydantic>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
rich>=10.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6 # For file uploads
httpx>=0.28.1,<0.29
prometheus_client
y-py>=0.6.0
redis>=4.6.0
rtree>=1.2.0
openai-agents>=0.0.11
structlog